# marketplace/openapi.yaml
openapi: 3.0.3
info:
  title: marketplace
  description: >
    Marketplace vendeur multi-clubs de l'écosystème `foot`. Trois surfaces :
    `/auth` (public), routes vendeur (`SellerJwtGuard`), et routes internes
    service-à-service (`ServiceAuthGuard`, parfois combiné à
    `AllowedApplicationsGuard` qui restreint à une liste explicite
    d'applications appelantes — voir moderation/payouts internes).

    TASK-P0-019 (todo.md) : document écrit à la main à partir du code réel
    des contrôleurs — couvre l'intégralité des routes et leur modèle
    d'autorisation (le problème central : "APIs implicites, changement
    cassant sans notification"), avec un niveau de détail plus léger sur
    les schémas de requête/réponse que payments/notifications vu le
    volume de cette API (~65 routes) — à enrichir DTO par DTO au besoin
    plutôt que reporté entièrement. Pas de génération automatique
    (`@nestjs/swagger`) branchée sur le build, voir la note dans todo.md.
  version: "1.0.0"
servers:
  - url: /
    description: Base URL configurée par déploiement (variable PORT)
paths:
...
  /sellers/me:
    get:
      operationId: getMySellerProfile
      summary: Profil du vendeur connecté
      tags: [sellers]
      security: [{ SellerJwt: [] }]
      responses:
        "200": { description: OK }
    patch:
      operationId: updateMySellerProfile
      summary: Met à jour le profil du vendeur connecté
      tags: [sellers]
      security: [{ SellerJwt: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "200": { description: OK }
  /sellers:
    get:
      operationId: listSellers
      summary: Liste tous les vendeurs (service-à-service)
      tags: [sellers]
      security: [{ ServiceApiKey: [] }]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200": { description: OK }
  /sellers/{id}:
    get:
      operationId: getSeller
      summary: Détail d'un vendeur (service-à-service)
      tags: [sellers]
      security: [{ ServiceApiKey: [] }]
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      responses:
        "200": { description: OK }
        "404": { description: Introuvable }
  /sellers/{id}/status:
    patch:
      operationId: updateSellerStatus
      summary: Change le statut de modération d'un vendeur (applications autorisées uniquement)
      tags: [sellers]
      security: [{ ServiceApiKey: [] }]
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [PENDING, APPROVED, SUSPENDED, REJECTED] }
      responses:
        "200": { description: OK }
        "403": { description: Application appelante non autorisée (AllowedApplicationsGuard) }
...
components:
  securitySchemes:
    SellerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT propre au vendeur, émis par POST /auth/login (pas le JWT sso).
    ServiceApiKey:
      type: apiKey
      in: header
      name: x-api-key
      description: >
        Clé de service par application appelante (ServiceAuthGuard). Sur
        certaines routes (annotées ci-dessus), en plus restreint à une
        liste explicite d'applications autorisées (AllowedApplicationsGuard).
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...